Create students.sql
This is an assignment for an SQL bootcamp from Lets Upgrade. In this CRUD operations are performed according to the questions mentioned in the assignment. 

Create a table called "students" with columns "id" (integer, primary key, auto-increment), "name" (text), "age" (integer), and "gender" (text).

Insert three rows of data into the "students" table.
Retrieve all rows from the "students" table.
Update the name of the student with id = 2 to "Janet".
Delete the row with id = 3 from the "students" table.
Retrieve all rows from the "students" table again to verify that the changes have been made.